+++ Changelog.md
@@ -1,4 +1,10 @@
-﻿1.0.0: 14.05.2021
+﻿1.1.0: 14.05.2021
+	-List is sorted by name/manufactorer
+	-Names will be trimmed before saving
+	-If there is no Rater, show Rater Change Fragment
+	-Min Rating is -1, will delete Rating
+
+1.0.0: 14.05.2021
 	-Initial Commit
 	-Main List finished
 	-Can add,remove,rate